Refactor: moved all the logic from controller into music service
